+++ Back/src/main/java/com/toogoodtogo/application/shop/product/ProductService.java
@@ -2,6 +2,7 @@ package com.toogoodtogo.application.shop.product;
 
 import com.toogoodtogo.application.S3Uploader;
 import com.toogoodtogo.application.UploadFileConverter;
+import com.toogoodtogo.domain.security.exceptions.CAccessDeniedException;
 import com.toogoodtogo.domain.shop.product.exceptions.CProductNotFoundException;
 import com.toogoodtogo.domain.shop.exceptions.CShopNotFoundException;
 import com.toogoodtogo.advice.exception.CValidCheckException;
@@ -10,22 +11,23 @@ import com.toogoodtogo.domain.shop.ShopRepository;
 import com.toogoodtogo.domain.shop.product.Product;
 import com.toogoodtogo.domain.shop.product.ProductRepository;
 import com.toogoodtogo.domain.shop.product.ProductRepositorySupport;
-import com.toogoodtogo.web.shops.products.dto.AddProductRequest;
-import com.toogoodtogo.web.shops.products.dto.ProductDto;
-import com.toogoodtogo.web.shops.products.dto.ProductTmp;
-import com.toogoodtogo.web.shops.products.dto.UpdateProductRequest;
+import com.toogoodtogo.web.shops.products.dto.*;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.nio.file.AccessDeniedException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class ProductService implements ProductUseCase {
     private final ProductRepository productRepository;
     private final ProductRepositorySupport productRepositorySupport;
@@ -62,12 +64,16 @@ public class ProductService implements ProductUseCase {
         if (filePath.equals("default.png")) fileName = "default.png"; // 기본 이미지
         else fileName = s3Uploader.upload(file, filePath); // 최종 파일 경로 및 파일 업로드
 
+        // 해당 shop 의 product 갯수
+        Long productPriority = productRepository.countProductByShopId(shopId);
+
         Product new_product = Product.builder()
                 .shop(shop)
                 .name(request.getName())
                 .price(request.getPrice())
                 .discountedPrice(request.getDiscountedPrice())
                 .image(fileName)
+                .priority(productPriority)
                 .build();
         return new ProductDto(productRepository.save(new_product));
 //        return ProductDto.builder().product(productRepository.save(new_product)).build();
@@ -77,15 +83,20 @@ public class ProductService implements ProductUseCase {
 //                .product(productRepository.save(request.toEntity(shop))).build();
     }
 
+    private boolean checkAccessOfShop(Long managerId, Long shopId) {
+        return shopRepository.findById(shopId).orElseThrow(CShopNotFoundException::new).getUser().getId().equals(managerId);
+    }
+
     @Transactional
-    public ProductDto updateProduct(Long managerId, Long productId, MultipartFile file, UpdateProductRequest request) throws IOException {
-        Product modifiedProduct = productRepository.findByUserIdAndId(managerId, productId).orElseThrow(CProductNotFoundException::new);
+    public ProductDto updateProduct(Long managerId, Long shopId, Long productId, MultipartFile file, UpdateProductRequest request) throws IOException {
+        // 로그인한 유저가 해당 shop 에 대해 권한 가졌는지 체크
+        if (!checkAccessOfShop(managerId, shopId)) throw new CAccessDeniedException();
+        Product modifiedProduct = productRepository.findByShopIdAndId(shopId, productId).orElseThrow(CProductNotFoundException::new);
         String filePath;
         String fileName;
         if (file.isEmpty()) { // 넘어온 사진이 없으면
             fileName = modifiedProduct.getImage(); // 기존 이미지
-        }
-        else { // 넘어온 사진이 있으면
+        } else { // 넘어온 사진이 있으면
             filePath = uploadFileConverter.parseFileInfo(file, "productsImage", modifiedProduct.getShop().getId());
             fileName = s3Uploader.updateS3(file, modifiedProduct.getImage(), filePath);
         }
@@ -99,8 +110,42 @@ public class ProductService implements ProductUseCase {
     }
 
     @Transactional
-    public String deleteProduct(Long managerId, Long productId) {
-        Product deleteProduct = productRepository.findByUserIdAndId(managerId, productId).orElseThrow(CProductNotFoundException::new);
+    public List<ProductDto> updateProductPriority(Long managerId, Long shopId, Long productId, UpdateProductPriorityRequest request) {
+        // 로그인한 유저가 해당 shop 에 대해 권한 가졌는지 체크
+        if (!checkAccessOfShop(managerId, shopId)) throw new CAccessDeniedException();
+        Product modifiedProduct = productRepository.findByShopIdAndId(shopId, productId).orElseThrow(CProductNotFoundException::new);;
+        // 해당 shop 의 priority 순으로 정렬된 product 들.
+        List<Product> products = productRepositorySupport.findProductsPerShopSortByPriority(shopId);
+        // 해당 shop 의 product 들 불러서 priority 로 정렬. prior 바뀌면 영향 가는 것들 순회하면서 +1 혹은 -1
+        Long beforePriority = modifiedProduct.getPriority();
+        Long afterPriority = request.getPriority();
+        // 바꾸려는 product 의 우선순위가 더 낮을 때 / 숫자가 더 클 때 / 앞으로 이동할 때
+        if (beforePriority > afterPriority) {
+            for (Product product : products) {
+                if (afterPriority <= product.getPriority() && product.getPriority() < beforePriority) {
+                    product.updatePriority(product.getPriority() + 1);
+                }
+            }
+            modifiedProduct.updatePriority(request.getPriority());
+        }
+        // 바꾸려는 product 의 우선순위가 더 높을 때 / 숫자가 더 낮을 때 / 뒤로 이동할 때
+        else if (beforePriority < afterPriority) {
+            for (Product product : products) {
+                if (beforePriority < product.getPriority() && product.getPriority() < afterPriority) {
+                    product.updatePriority(product.getPriority() - 1);
+                }
+            }
+            modifiedProduct.updatePriority(request.getPriority()-1);
+        }
+        return productRepositorySupport.findProductsPerShopSortByPriority(shopId)
+                .stream().map(ProductDto::new).collect(Collectors.toList());
+    }
+
+    @Transactional
+    public String deleteProduct(Long managerId, Long shopId, Long productId) {
+        // 로그인한 유저가 해당 shop 에 대해 권한 가졌는지 체크
+        if (!checkAccessOfShop(managerId, shopId)) throw new CAccessDeniedException();
+        Product deleteProduct = productRepository.findByShopIdAndId(shopId, productId).orElseThrow(CProductNotFoundException::new);
         // 기본 이미지가 아니면 S3에서 이미지 삭제
         if (!deleteProduct.getImage().equals("default.png")) s3Uploader.deleteFileS3(deleteProduct.getImage());
         productRepository.deleteById(productId);
@@ -113,12 +158,13 @@ public class ProductService implements ProductUseCase {
     }
 
     @Transactional(readOnly = true)
-    public List<List<ProductDto>> productsPerCategory(String category) {
-        return productRepositorySupport.productsPerCategory(category);
+    public List<ProductDto> productsPerCategory(String category, String method) {
+        return productRepositorySupport.productsPerCategory2(category, method);
     }
 
     @Transactional(readOnly = true)
-    public List<ProductDto> sortProductsPerShop(Long shopId, String method) {
-        return productRepositorySupport.sortProductsPerShop(shopId, method);
+    public List<ProductDto> findProductsPerShopSortByPriority(Long shopId) {
+        return productRepositorySupport.findProductsPerShopSortByPriority(shopId)
+                .stream().map(ProductDto::new).collect(Collectors.toList());
     }
 }
